Working version w/ cast widget Now, need to merge w/ recent changes from main repo

+++ core/src/main/java/com/capitalone/dashboard/model/SprintEstimate.java
@@ -1,56 +1,32 @@
 package com.capitalone.dashboard.model;
 
 public class SprintEstimate {
-	private int openEstimate;
-	private int inProgressEstimate;
-	private int completeEstimate;
-	private int totalEstimate;
-	/**
-	 * @return the openEstimate
-	 */
-	public int getOpenEstimate() {
+	private double openEstimate;
+	private double inProgressEstimate;
+	private double completeEstimate;
+	private double totalEstimate;
+	public double getOpenEstimate() {
 		return openEstimate;
 	}
-	/**
-	 * @param openEstimate the openEstimate to set
-	 */
-	public void setOpenEstimate(int openEstimate) {
+	public void setOpenEstimate(double openEstimate) {
 		this.openEstimate = openEstimate;
 	}
-	/**
-	 * @return the inProgressEstimate
-	 */
-	public int getInProgressEstimate() {
+	public double getInProgressEstimate() {
 		return inProgressEstimate;
 	}
-	/**
-	 * @param inProgressEstimate the inProgressEstimate to set
-	 */
-	public void setInProgressEstimate(int inProgressEstimate) {
+	public void setInProgressEstimate(double inProgressEstimate) {
 		this.inProgressEstimate = inProgressEstimate;
 	}
-	/**
-	 * @return the completeEstimate
-	 */
-	public int getCompleteEstimate() {
+	public double getCompleteEstimate() {
 		return completeEstimate;
 	}
-	/**
-	 * @param completeEstimate the completeEstimate to set
-	 */
-	public void setCompleteEstimate(int completeEstimate) {
+	public void setCompleteEstimate(double completeEstimate) {
 		this.completeEstimate = completeEstimate;
 	}
-	/**
-	 * @return the totalEstimate
-	 */
-	public int getTotalEstimate() {
+	public double getTotalEstimate() {
 		return totalEstimate;
 	}
-	/**
-	 * @param totalEstimate the totalEstimate to set
-	 */
-	public void setTotalEstimate(int totalEstimate) {
+	public void setTotalEstimate(double totalEstimate) {
 		this.totalEstimate = totalEstimate;
 	}
 }
